refactor: moved LookupEnum to dedicated file to fix circular imports
