+++ src/modalities/config/lookup_enum.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class LookupEnum(Enum):
+    @classmethod
+    def _missing_(cls, value: str) -> type:
+        """constructs Enum by member name, if not constructable by value"""
+        return cls.__dict__[value]
